pyproject.toml
[project]
name = "input-capture-api"
description = "A simple library for recording raw HID input events."
dynamic = ["version"]
readme = "README.md"
authors = [
    { name = "osoken" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.120.1",
    "hid-interceptor",
    "hid-recorder",
    "python-ulid>=3.1.0",
    "uvicorn[standard]>=0.38.0",
]

[build-system]
requires = ["setuptools>=70.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "input_capture_api.__version__"}

[tool.uv.sources]
hid-interceptor = { git = "https://github.com/osoekawaitlab/hid-interceptor" }
hid-recorder = { git = "https://github.com/osoekawaitlab/hid-recorder" }

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "nox>=2025.10.16",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.2",
]

[tool.ruff]
indent-width = 4
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # trailing-comma-missing (conflicts with formatter)
    "ISC001",  # single-line-implicit-string-concatenation (conflicts with formatter)
    "D203",    # one-blank-line-before-class (conflicts with D211)
    "D213",    # multi-line-summary-second-line (conflicts with D212)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert-used (pytest uses asserts)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests", "src"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--doctest-modules",
]
markers = [
]
